refactor: updated handlers import
